Session 10 - Beginnings of incremental merge
* Cleaned up some compilation errors
* Created the HanoiDB struct containing the config and the nursery and
levels
* Return commands from individual writes that indicate whether to
promote a file to the next level or to trigger an incremental merge
* Implement promoting nursery data to the top level
* Stub out handling incremental merge

+++ src/merger.rs
@@ -25,7 +25,7 @@ impl std::fmt::Debug for Merger {
 impl Merger {
     pub(crate) fn new(
         path: impl AsRef<Path>,
-        level: usize,
+        level: u32,
         a_tree: &Tree,
         b_tree: &Tree,
     ) -> Result<Self> {
